Add files via upload
These files were uploaded and not pushed using a linux computer because during the time that they were modified (9/8/18-9/10/18) the GWU jupyter notebook was unaccessible. Because of this, Anaconda was downloaded to my laptop, the files were recreated, and then uploaded to github. My laptop is a windows laptop and the jupyter notebook that was downloaded is a copy of my computer, and is a local host. Because of this, the files were uploaded to github and not pushed via the linux terminal. If my laptop had been a linux machine or if the jupyter notebook had been accessible from a linux computer, the files could have been pushed to github via the linux terminal.